Implementa toogle-feature para a remoção do segmento Outros

+++ plugins/SettingsPa/db-updates.php
@@ -10,6 +10,12 @@ return [
         $app = App::i();
         $em = $app->em;
         $conn = $em->getConnection();
-        $conn->executeQuery("delete from term where taxonomy = 'segmento' and term = 'Outros'");
+        if($app->config['plugins']['SettingsPa']['config']['remove_other_segment']) {
+            $conn->executeQuery("delete from term where taxonomy = 'segmento' and term = 'Outros'");
+            $app->log->info("Todos os segmentos foram removidos");
+        } else {
+            $app->log->info("Segmento Outros não removido, configuração REMOVE_OTHER_SEGMENT está desativada (Roda apenas no Pará) Olhar plugins SettingsPa para ver implementação");
+            return false;
+        }
     }
 ];
